# Shopify Service/static/swagger.yaml
openapi: 3.0.3
info:
  title: Shopify App
  version: 1.0.0
  contact: {}
servers:
  - url: http://127.0.0.1
paths:
...
  /shopify_rest/update_product:
    post:
      summary: Update Product
      description: Update Product
      operationId: updateProduct
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: Now flaunt your wrist with out brand new watch
                id:
                  type: string
                  example: '8376743887126'
                store_name:
                  type: string
                  example: computelibstore
                title:
                  type: string
                  example: The new Watch Class
            examples:
              Update Product:
                value:
                  description: Now flaunt your wrist with out brand new watch
                  id: '8376743887126'
                  store_name: computelibstore
                  title: The new Watch Class
      responses:
        '200':
          description: ''
...